requirements.txt
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
joblib==1.3.2
pyyaml==6.0.1
python-dotenv==1.0.0
astropy==5.3.4
lightkurve==2.4.2
